+++ Microsoft.Azure.Cosmos/src/Telemetry/OpenTelemetry/OpenTelemetryResponse.cs
@@ -5,16 +5,20 @@
 namespace Microsoft.Azure.Cosmos
 {
     using System.Net;
+    using Microsoft.Azure.Cosmos.Diagnostics;
     using Microsoft.Azure.Documents;
 
-    /// <summary>
-    /// IOpenTelemetryResponse
-    /// </summary>
-    internal class OpenTelemetryResponse
+    internal sealed class OpenTelemetryResponse
     {
         public OpenTelemetryResponse(ResponseMessage message)
         {
             this.StatusCode = message.StatusCode;
+            this.RequestCharge = message.Headers?.RequestCharge;
+            this.RequestContentLength = message.RequestMessage.Headers?.ContentLength;
+            this.ResponseContentLength = message.Headers?.ContentLength;
+            this.ContainerName = message.RequestMessage.ContainerId;
+            this.Diagnostics = message.Diagnostics;
+            //TODO: ItemCount needs to be added
         }
 
         /// <summary>
@@ -25,17 +29,17 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// RequestCharge
         /// </summary>
-        public double RequestCharge { get; }
+        public double? RequestCharge { get; }
 
         /// <summary>
         /// RequestLength
         /// </summary>
-        public long? RequestContentLength { get; }
+        public string RequestContentLength { get; }
 
         /// <summary>
         /// ResponseLength
         /// </summary>
-        public long? ResponseContentLength { get; }
+        public string ResponseContentLength { get; }
 
         /// <summary>
         /// ContainerName
@@ -46,6 +50,11 @@ namespace Microsoft.Azure.Cosmos
         /// ItemCount
         /// </summary>
         public string ItemCount { get; }
+
+        /// <summary>
+        /// ItemCount
+        /// </summary>
+        public CosmosDiagnostics Diagnostics { get; }
 
     }
 }
